+++ trunk/mov/src/org/mov/parser/expression/IfExpression.java
@@ -43,7 +43,7 @@ public class IfExpression extends TernaryExpression {
 	super(arg1, arg2, arg3);
     }
 
-    public float evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) 
+    public double evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) 
 	throws EvaluationException {
 
 	// if(...) then
